# apps/api/internal/httpapi/me_push.go
package httpapi

import (
	"errors"
	"net/http"
	"net/url"
	"strings"

	"github.com/openclaw/clickclack/apps/api/internal/store"
	"github.com/openclaw/clickclack/apps/api/internal/webpush"
)

// webPushEnabled reports whether the server can deliver web push. Without a
// configured key pair the endpoints answer as if the feature does not exist,
// and no subscription row is ever written.
func (s *Server) webPushEnabled() bool {
	return s.webPushNotifier != nil && s.webPushPublicKey != ""
}

func (s *Server) getMyPush(w http.ResponseWriter, r *http.Request) {
	act, err := s.currentActor(r)
	if err != nil {
		writeError(w, http.StatusUnauthorized, err)
		return
	}
	if err := act.requireScope("profile:read"); err != nil {
		writeError(w, http.StatusForbidden, err)
		return
	}
	if !s.webPushEnabled() {
		writeJSON(w, http.StatusOK, map[string]any{
			"enabled":          false,
			"vapid_public_key": "",
			"subscriptions":    []store.PushSubscription{},
		})
		return
	}
	subscriptions, err := s.store.ListPushSubscriptions(r.Context(), act.user.ID)
	if err != nil {
		writeStoreError(w, err)
		return
	}
	writeJSON(w, http.StatusOK, map[string]any{
		"enabled":          true,
		"vapid_public_key": s.webPushPublicKey,
		"subscriptions":    subscriptions,
	})
}

func (s *Server) putMyPushSubscription(w http.ResponseWriter, r *http.Request) {
	act, ok := s.requirePushActor(w, r)
	if !ok {
		return
	}
	var body struct {
		Endpoint string `json:"endpoint"`
		Keys     struct {
			P256dh string `json:"p256dh"`
			Auth   string `json:"auth"`
		} `json:"keys"`
		UserAgent string `json:"user_agent"`
	}
	if err := readJSON(w, r, &body); err != nil {
		writeError(w, http.StatusBadRequest, err)
		return
	}
	if err := validatePushEndpoint(body.Endpoint); err != nil {
		writeError(w, http.StatusBadRequest, err)
		return
	}
	if err := webpush.ValidateSubscriptionKeys(body.Keys.P256dh, body.Keys.Auth); err != nil {
		writeError(w, http.StatusBadRequest, err)
		return
	}
	subscription, err := s.store.UpsertPushSubscription(r.Context(), store.PushSubscriptionInput{
		UserID:       act.user.ID,
		Endpoint:     body.Endpoint,
		P256dh:       body.Keys.P256dh,
		Auth:         body.Keys.Auth,
		UserAgent:    body.UserAgent,
		SessionToken: act.sessionToken,
	})
	writeResult(w, map[string]any{"subscription": subscription}, err)
}

func (s *Server) deleteMyPushSubscription(w http.ResponseWriter, r *http.Request) {
	act, ok := s.requirePushActor(w, r)
	if !ok {
		return
	}
	var body struct {
		Endpoint string `json:"endpoint"`
	}
	if err := readJSON(w, r, &body); err != nil {
		writeError(w, http.StatusBadRequest, err)
		return
	}
	endpoint := strings.TrimSpace(body.Endpoint)
	if endpoint == "" {
		writeError(w, http.StatusBadRequest, errors.New("endpoint is required"))
		return
	}
	if err := s.store.DeletePushSubscription(r.Context(), act.user.ID, endpoint); err != nil {
		writeStoreError(w, err)
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

// requirePushActor resolves the signed-in human behind a subscription change.
// Bot tokens have no device to notify, and a server without keys has no
// subscriptions at all.
func (s *Server) requirePushActor(w http.ResponseWriter, r *http.Request) (actor, bool) {
	act, err := s.currentActor(r)
	if err != nil {
		writeError(w, http.StatusUnauthorized, err)
		return actor{}, false
	}
	if act.botTokenID != "" {
		writeError(w, http.StatusForbidden, errors.New("bot tokens cannot register push subscriptions"))
		return actor{}, false
	}
	if !s.webPushEnabled() {
		writeError(w, http.StatusNotFound, errors.New("web push is not configured"))
		return actor{}, false
	}
	return act, true
}

func validatePushEndpoint(endpoint string) error {
	endpoint = strings.TrimSpace(endpoint)
	if endpoint == "" {
		return errors.New("endpoint is required")
	}
	parsed, err := url.Parse(endpoint)
	if err != nil || parsed.Host == "" || parsed.User != nil {
		return errors.New("endpoint must be an absolute push service URL")
	}
	return validatePushEndpointHost(parsed.Scheme, parsed.Hostname())
}